修改log

+++ game_analyze/src/com/gamecenter/mapper/OpOssQlzCreateroleLogMapper.java
@@ -6,25 +6,25 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface OpOssQlzCreateroleLogMapper {
-	int countByExample(OpOssQlzCreateroleLogExample example);
-	
-	int deleteByExample(OpOssQlzCreateroleLogExample example);
-	
-	int deleteByPrimaryKey(Integer did);
-	
-	int insert(OpOssQlzCreateroleLog record);
-	
-	int insertSelective(OpOssQlzCreateroleLog record);
-	
-	List<OpOssQlzCreateroleLog> selectByExample(OpOssQlzCreateroleLogExample example);
-	
-	OpOssQlzCreateroleLog selectByPrimaryKey(Integer did);
-	
-	int updateByExampleSelective(@Param("record") OpOssQlzCreateroleLog record, @Param("example") OpOssQlzCreateroleLogExample example);
-	
-	int updateByExample(@Param("record") OpOssQlzCreateroleLog record, @Param("example") OpOssQlzCreateroleLogExample example);
-	
-	int updateByPrimaryKeySelective(OpOssQlzCreateroleLog record);
-	
-	int updateByPrimaryKey(OpOssQlzCreateroleLog record);
+    long countByExample(OpOssQlzCreateroleLogExample example);
+
+    int deleteByExample(OpOssQlzCreateroleLogExample example);
+
+    int deleteByPrimaryKey(Integer did);
+
+    int insert(OpOssQlzCreateroleLog record);
+
+    int insertSelective(OpOssQlzCreateroleLog record);
+
+    List<OpOssQlzCreateroleLog> selectByExample(OpOssQlzCreateroleLogExample example);
+
+    OpOssQlzCreateroleLog selectByPrimaryKey(Integer did);
+
+    int updateByExampleSelective(@Param("record") OpOssQlzCreateroleLog record, @Param("example") OpOssQlzCreateroleLogExample example);
+
+    int updateByExample(@Param("record") OpOssQlzCreateroleLog record, @Param("example") OpOssQlzCreateroleLogExample example);
+
+    int updateByPrimaryKeySelective(OpOssQlzCreateroleLog record);
+
+    int updateByPrimaryKey(OpOssQlzCreateroleLog record);
 }
